+++ google_cloud_automlops/utils/utils.py
@@ -462,6 +462,7 @@ def get_required_apis(defaults: dict) -> Set:
         set: The set of required APIs
     """
     required_apis = set([
+        'cloudbuild.googleapis.com',
         'cloudresourcemanager.googleapis.com',
         'compute.googleapis.com',
         'iamcredentials.googleapis.com',
@@ -472,8 +473,8 @@ def get_required_apis(defaults: dict) -> Set:
         required_apis.add('aiplatform.googleapis.com')
     if defaults['gcp']['artifact_repo_type'] == ArtifactRepository.ARTIFACT_REGISTRY.value:
         required_apis.add('artifactregistry.googleapis.com')
-    if defaults['tooling']['deployment_framework'] == Deployer.CLOUDBUILD.value:
-        required_apis.add('cloudbuild.googleapis.com')
+    # if defaults['tooling']['deployment_framework'] == Deployer.CLOUDBUILD.value:
+    #     required_apis.add('cloudbuild.googleapis.com')
     if defaults['gcp']['schedule_pattern'] != DEFAULT_SCHEDULE_PATTERN:
         required_apis.add('cloudscheduler.googleapis.com')
     if defaults['gcp']['pipeline_job_submission_service_type'] == PipelineJobSubmitter.CLOUD_RUN.value: